models call on contract
